requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
idna==2.9
inflection==0.4.0
itypes==1.1.0
Jinja2==2.11.2
jsonapi==0.9.9
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.1.1
psycopg2==2.8.5
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
